requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.19.0
tokenizers>=0.19.0
seqeval>=1.2.2
gliner>=0.2.0
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
openai>=1.30.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
seaborn>=0.13.0
pandas>=2.2.0
pyyaml>=6.0.0
structlog>=24.1.0
tqdm>=4.66.0
ruff>=0.4.0
onnxruntime>=1.17.0
onnxscript>=0.1.0
sklearn-crfsuite>=0.3.6
google-generativeai>=0.5.0
pytest>=8.0.0
httpx>=0.27.0